Moved oath template to templates dir
