+++ api/src/middlewares/services/EmailService/templates/index.ts
@@ -1,0 +1,2 @@
+import oath from "./oath";
+export { oath };
