+++ api/src/middlewares/services/EmailService/templates/oath.ts
